requirements.txt
psycopg2-binary
sqlalchemy
schedule
python-dotenv